Fix bugs in the Plane class
From Glenn Burkhardt:

There's a bug in the Plane class in George Vanecek's "Spatial
Partitioning of a Polygon by a Plane".

In 'plane.cc', at line 14, there's an obvious array out of bounds
reference:

for( Index i = 0; i < nPoints; ++i ) {
avrPnt += points[i-1];

Clearly, we don't want to reference 'points[-1]', which will happen on
the first iteration.

Also, in 'vector.h', the normalize method doesn't normalize the vector,
since it divides by the square of the magnitude, not the magnitude.
Probably, that doesn't matter for this algorithm.

The computation of the ‘eps’ value when computing the plane’s unit
vector was done prematurely.  It doesn’t matter for the 2D cases.  I was
trying to generate cut planes from a 2D line segment for testing.